Create App.js
This is the entry point of the app and contains the main component of the app.

This code sets up the basic structure of the app and contains the NavigationContainer and StackNavigator from the @react-navigation library for managing navigation between screens. It also sets up an authentication context with AuthContext and provides this context to the rest of the app using the AuthContext.Provider.

The getToken, storeToken, and removeToken functions are used for storing and retrieving a user token, which is used to maintain the user's authentication state.

The useEffect hook is used to check for a stored user token when the app is launched and to set the initial state of the app accordingly.

The AuthStack and Navigation components are not defined here, but will be defined in separate files. The AuthStack component will contain the login and registration screens, and the Navigation component will contain the main navigation flow of the app.